@@ -53,7 +53,7 @@ class _AppAudioLanguageState extends ConsumerState<AudioLanguageScreen> {
           child: SafeArea(
             child: Padding(
               padding: const EdgeInsets.all(16),
-              child: OptionList(
+              child: OptionList<String>(
                 optionData: languages.entries.map((e) => Option(id: e.key, title: e.value.nativeName, subTitle: e.value.englishName)).toList(),
                 currentSelection: selected,
                 backgroundColor: Colors.transparent,
